+++ src/auth/auth.module.ts
@@ -5,13 +5,12 @@ import { PassportModule } from '@nestjs/passport';
 import { UserModule } from '../user/user.module';
 import { AuthController } from './auth.controller';
 import { AuthService } from './auth.service';
-import { GoogleStrategy } from './strategies/google/google.strategy';
 import { JwtStrategy } from './strategies/jwt/jwt.strategy';
 import { LocalStrategy } from './strategies/local/local.strategy';
 
 @Module({
     controllers: [AuthController],
-    providers: [AuthService, LocalStrategy, JwtStrategy, GoogleStrategy],
+    providers: [AuthService, LocalStrategy, JwtStrategy],
     imports: [
         UserModule,
         PassportModule,
