+++ src/auth/dto/google-token.dto.ts
@@ -1,0 +1,7 @@
+import { IsNotEmpty, IsString } from 'class-validator';
+
+export class GoogleToken {
+    @IsString()
+    @IsNotEmpty()
+    token: string;
+}
