--- src/auth/strategies/google/google-auth.guard.ts
@@ -1,3 +1,0 @@
-import { AuthGuard } from '@nestjs/passport'
-
-export class GoogleAuthGuard extends AuthGuard('google') {}
