+++ test/auth.e2e-spec.ts
@@ -6,12 +6,11 @@ import { Test, TestingModule } from '@nestjs/testing';
 import { Connection, Model } from 'mongoose';
 import * as request from 'supertest';
 import { AuthModule } from '../src/auth/auth.module';
-import { GoogleAuthGuard } from '../src/auth/strategies/google/google-auth.guard';
+import { AuthService } from '../src/auth/auth.service';
 import { UserDocument } from '../src/user/entities/user.entity';
 import { UserStatus } from '../src/user/enums/status.enum';
 import { UserModule } from '../src/user/user.module';
 import { ThirdPartyGuardMock } from './helpers/fake-provider-strategy';
-import { ProviderGuardFaker } from './helpers/fake-third-party-guard';
 import {
     closeInMongodConnection,
     rootMongooseTestModule
@@ -23,6 +22,7 @@ describe('AuthMdoule (e2e)', () => {
     let app: INestApplication;
     let connection: Connection;
     let userModel: Model<UserDocument>;
+    let authService: AuthService;
 
     beforeAll(async () => {
         const module: TestingModule = await Test.createTestingModule({
@@ -36,12 +36,10 @@ describe('AuthMdoule (e2e)', () => {
                 })
             ],
             providers: [ThirdPartyGuardMock]
-        })
-            .overrideGuard(GoogleAuthGuard) // Overwrite guards with mocks that don´t rely on external APIs
-            .useClass(ProviderGuardFaker)
-            .compile();
+        }).compile();
 
         connection = await module.get(getConnectionToken());
+        authService = module.get<AuthService>(AuthService);
         userModel = connection.model('User');
         app = module.createNestApplication();
         app.useGlobalPipes(new ValidationPipe({ whitelist: true }));
@@ -97,83 +95,6 @@ describe('AuthMdoule (e2e)', () => {
                         password: 'mensa essen'
                     })
                     .expect(HttpStatus.CONFLICT);
-            });
-        });
-
-        describe('/auth/(third-party-provider) (GET)', () => {
-            it('/auth/google should create user', async () => {
-                // send data that normally is provided by guard
-                await request(app.getHttpServer())
-                    .post('/auth/google')
-                    .send({
-                        user: {
-                            username: 'googleman',
-                            email: 'googleuser@google.com',
-                            provider: 'google'
-                        }
-                    })
-                    .expect(HttpStatus.CREATED);
-                expect(await (await userModel.find()).length).toBe(1);
-            });
-
-            it('/auth/google can be used for login (given user has provider)', async () => {
-                // add provide to test user
-                let user = await getTestUser();
-                user = { ...user, provider: 'google' };
-                await userModel.create(user);
-                // send data that normally is provided by guard
-                await request(app.getHttpServer())
-                    .post('/auth/google')
-                    .send({
-                        user: {
-                            username: 'mock',
-                            email: 'mock@mock.mock',
-                            provider: 'google'
-                        }
-                    })
-                    .expect(HttpStatus.CREATED);
-                expect(await (await userModel.find()).length).toBe(1);
-            });
-
-            it('/auth/google should throw error on duplicate', async () => {
-                // send data that normally is provided by guard
-                await userModel.create(await getTestUser());
-                await request(app.getHttpServer())
-                    .post('/auth/google')
-                    .send({
-                        user: {
-                            username: 'mock',
-                            email: 'mock@mock.mock',
-                            provider: 'google'
-                        }
-                    })
-                    .expect(HttpStatus.CONFLICT);
-                expect(await (await userModel.find()).length).toBe(1);
-            });
-
-            it('/auth/google should throw error on duplicate username', async () => {
-                // send data that normally is provided by guard
-                await userModel.create(await getTestUser());
-                await request(app.getHttpServer())
-                    .post('/auth/google')
-                    .send({
-                        user: {
-                            username: 'mock',
-                            email: 'not@mock.mock',
-                            provider: 'google'
-                        }
-                    })
-                    .expect(HttpStatus.INTERNAL_SERVER_ERROR);
-                expect(await (await userModel.find()).length).toBe(1);
-            });
-
-            it('/auth/google should fail without values', async () => {
-                // send data that normally is provided by guard
-                await userModel.create(await getTestUser());
-                await request(app.getHttpServer())
-                    .post('/auth/google')
-                    .send({})
-                    .expect(HttpStatus.UNAUTHORIZED);
             });
         });
